implement Mesh, delete Solid
