+++ src/geometry.cpp
@@ -1,97 +1,185 @@
 #include <graphics/geometry.h>
 #include <glm/gtx/string_cast.hpp>
+
 using namespace std;
 using namespace graphics;
 
-// Facet
-//--------------------------------------------------------------------------------
-Facet::Facet(GLuint a, GLuint b, GLuint c)
-: a(a), b(b), c(c) {
-  indices[0] = &a;
-  indices[1] = &b;
-  indices[2] = &c;
-}
-
-GLuint Facet::operator[](int idx) {
-  return *(indices[idx]);
-}
-
 // Geometry
 //--------------------------------------------------------------------------------
-Geometry::Geometry(std::vector<Facet> _facets,
-                   const std::vector<glm::vec3>& _positions,
-                   bool smooth,
-                   std::vector<glm::vec3> _normals,
-                   std::vector<glm::vec3> _tex_coords) :
-facets(std::move(_facets)), positions(std::move(_positions)),
-normals(std::move(_normals)), tex_coords(std::move(_tex_coords)) {
-  create_indices_from_facets();
-  if(_normals.empty()) {
-    compute_smooth_normals();
-  }
+Geometry::Geometry(const std::vector<std::shared_ptr<Mesh>>& _meshes)
+: meshes(std::move(_meshes)) {
+
 }
 
-Geometry::Geometry(const std::vector<glm::vec3>& _positions,
-                   std::vector<GLuint> _indices,
-                   bool smooth):
-positions(std::move(_positions)), indices(_indices) {
-  if(_indices.empty()) {
-    for(int i = 0; i < positions.size()/3; i++)
-      facets.emplace_back(i*3, i*3 + 1, i*3 + 2);
-    create_indices_from_facets();
-  } else {
-    for(int i = 0; i < indices.size()/3; i++)
-      facets.emplace_back(indices[i*3], indices[i*3 + 1], indices[i*3 + 2]);
-  }
-
-  compute_smooth_normals();
+void Geometry::draw() {
+  for(auto mesh: meshes)
+    mesh->draw();
 }
 
-void Geometry::compute_smooth_normals() {
-  vector<vector<Facet*>> assoc(positions.size());
-  for(int i = 0; i < facets.size(); i++) {
-    assoc[facets[i].a].push_back( &(facets[i]) );
-    assoc[facets[i].b].push_back( &(facets[i]) );
-    assoc[facets[i].c].push_back( &(facets[i]) );
+std::shared_ptr<Geometry> Geometry::create_box(GLfloat x, GLfloat y, GLfloat z) {
+  x = x/2;
+  y = y/2;
+  z = z/2;
+
+  auto ftl = glm::vec3(-x,y,z);   //front top left
+  auto ftr = glm::vec3(x,y,z);    //front top right
+  auto fll = glm::vec3(-x,-y,z);  //front low left
+  auto flr = glm::vec3(x,-y,z);   //front low right
+  auto btl = glm::vec3(x,y,-z);   //back top left
+  auto btr = glm::vec3(-x,y,-z);  //back top right
+  auto bll = glm::vec3(x,-y,-z);  //back low left
+  auto blr = glm::vec3(-x,-y,-z);  //back low right
+
+  vector<vector<glm::vec3>> face_positions = {
+    { ftl, ftr, fll, flr }, //front
+    { btl, btr, bll, blr }, //back
+    { btr, ftl, blr, fll }, //left
+    { ftr, btl, flr, bll }, //right
+    { btr, btl, ftl, ftr }, //top
+    { fll, flr, blr, bll }  //bottom
+  };
+
+  std::vector<GLuint> indices = {0,2,1, 1,2,3};
+  vector<shared_ptr<Mesh>> meshes;
+  for(auto& positons: face_positions) {
+    auto mesh = make_shared<Mesh>(positons, indices);
+    meshes.push_back(mesh);
   }
-  normals = vector<glm::vec3>(positions.size());
-  for(int i = 0; i < assoc.size(); i++) {
-    glm::vec3 normal;
-    for(int j = 0; j < assoc[i].size(); j++) {
-      normal += get_facet_normal( *(assoc[i][j]) ) / float(assoc[i].size());
+
+  shared_ptr<Geometry> box = make_shared<Geometry>(meshes);
+  return box;
+}
+
+std::shared_ptr<Geometry> Geometry::create_flat_sphere(GLfloat r, int st, int sc) {
+  if(st < 2 || sc < 3)
+    throw runtime_error("cannot have < 2 stacks or < 3 slices");
+  std::vector<glm::vec3> positions;
+  std::vector<GLint> indices;
+
+  vector<vector<glm::vec3>> temp(st + 1, {});
+  //top section
+  temp.front().emplace_back(0,r,0);
+  //mid sections
+  float per_st_angle = 180.0 / st;
+  float per_sc_angle = 360.0 / sc;
+  for(int i = 1; i < st; i++) {
+    float st_angle = 90.0 - (per_st_angle * i);
+    float sc_r = r * cos(glm::radians(st_angle));
+    float y = r * sin(glm::radians(st_angle));
+    for(int j = 0; j < sc; j++) {
+      float sc_angle = per_sc_angle * j;
+      float x = sc_r * sin(glm::radians(sc_angle));
+      float z = sc_r * cos(glm::radians(sc_angle));
+
+      temp[i].emplace_back(x,y,z);
     }
-    normals[i] = normal;
   }
-}
+  //low section
+  temp.back().emplace_back(0,-r,0);
 
-const Geometry& Geometry::operator+=(const Geometry& r) {
-  auto max = std::max_element(indices.begin(), indices.end());
-  int offset = (max == indices.end()) ? 0 : *max + 1;
-  for(int i = 0; i < r.facets.size(); i++)
-    facets.emplace_back( r.facets[i].a + offset, r.facets[i].b + offset, r.facets[i].c + offset);
-  for(int i = 0; i < r.indices.size(); i++)
-    indices.emplace_back(r.indices[i] + offset);
-  positions.insert(positions.end(), r.positions.begin(), r.positions.end());
-  normals.insert(normals.end(), r.normals.begin(), r.normals.end());
-  tex_coords.insert(tex_coords.end(), r.tex_coords.begin(), r.tex_coords.end());
-  return *this;
-}
-
-void Geometry::create_indices_from_facets() {
-  indices.clear();
-  for(int i = 0; i < facets.size(); i++) {
-    indices.push_back(facets[i].a);
-    indices.push_back(facets[i].b);
-    indices.push_back(facets[i].c);
+  //assign vertices and indices
+  //top
+  for(int j = 0; j < temp[1].size(); j++) {
+    positions.push_back(temp[0][0]);
+    positions.push_back(temp[1][j]);
+    positions.push_back(temp[1][(j + 1) % temp[1].size()]);
   }
+
+  //mid
+  for(int i = 1; i < temp.size() - 2; i++) {
+    for(int j = 0; j < temp[i].size(); j++) {
+      positions.push_back(temp[i][j]);
+      positions.push_back(temp[i + 1][j]);
+      positions.push_back(temp[i + 1][(j + 1) % temp[i].size()]);
+
+      positions.push_back(temp[i][j]);
+      positions.push_back(temp[i + 1][(j + 1) % temp[i].size()]);
+      positions.push_back(temp[i][(j + 1) % temp[i].size()]);
+    }
+  }
+
+  //bottom
+  int pen_idx = temp.size() - 2;
+  for(int j = 0; j < temp[pen_idx].size(); j++) {
+    positions.push_back(temp[pen_idx][j]);
+    positions.push_back(temp.back()[0]);
+    positions.push_back(temp[pen_idx][(j + 1) % temp[pen_idx].size()]);
+  }
+
+  vector<shared_ptr<Mesh>> meshes = { make_shared<Mesh>(positions) };
+  shared_ptr<graphics::Geometry> sphere = make_shared<Geometry>(meshes);
+  return sphere;
 }
 
-glm::vec3 Geometry::get_facet_normal(const Facet& facet) {
-  glm::vec3 pa = positions[facet.a];
-  glm::vec3 pb = positions[facet.b];
-  glm::vec3 pc = positions[facet.c];
-  glm::vec3 normal = glm::normalize(glm::cross(pc - pb, pa - pb));
-  return normal;
+std::shared_ptr<Geometry> Geometry::create_smooth_sphere(GLfloat r, int st, int sc) {
+  if(st < 2 || sc < 3)
+    throw runtime_error("cannot have < 2 stacks or < 3 slices");
+  std::vector<glm::vec3> positions;
+  std::vector<GLint> indices;
+
+  vector<vector<int>> grid_indices(st + 1, {}); //position indices of grid points
+  int accum = 0;
+
+  //top section
+  grid_indices.front().push_back(accum++);
+  positions.emplace_back(0,r,0);
+  //mid sections
+  float per_st_angle = 180.0 / st;
+  float per_sc_angle = 360.0 / sc;
+  for(int i = 1; i < st; i++) {
+    float st_angle = 90.0 - (per_st_angle * i);
+    float sc_r = r * cos(glm::radians(st_angle));
+    float y = r * sin(glm::radians(st_angle));
+    for(int j = 0; j < sc; j++) {
+      float sc_angle = per_sc_angle * j;
+      float x = sc_r * sin(glm::radians(sc_angle));
+      float z = sc_r * cos(glm::radians(sc_angle));
+
+      positions.emplace_back(x,y,z);
+      grid_indices[i].push_back(accum++);
+    }
+  }
+  //low section
+  grid_indices.back().push_back(accum++);
+  positions.emplace_back(0,-r,0);
+
+  //assign vertices and indices
+  vector<Facet> facets;
+  //top
+  for(int j = 0; j < grid_indices[1].size(); j++) {
+    int a = grid_indices[0][0];
+    int b = grid_indices[1][j];
+    int c = grid_indices[1][(j + 1) % grid_indices[1].size()];
+    facets.emplace_back(a,b,c);
+  }
+
+  //mid
+  for(int i = 1; i < grid_indices.size() - 2; i++) {
+    for(int j = 0; j < grid_indices[i].size(); j++) {
+      int a = grid_indices[i][j];
+      int b = grid_indices[i + 1][j];
+      int c = grid_indices[i + 1][(j + 1) % grid_indices[i].size()];
+      facets.emplace_back(a,b,c);
+
+      int d = grid_indices[i][j];
+      int e = grid_indices[i + 1][(j + 1) % grid_indices[i].size()];
+      int f = grid_indices[i][(j + 1) % grid_indices[i].size()];
+      facets.emplace_back(d,e,f);
+    }
+  }
+
+  //bottom
+  int pen_idx = grid_indices.size() - 2;
+  for(int j = 0; j < grid_indices[pen_idx].size(); j++) {
+    int a = grid_indices[pen_idx][j];
+    int b = grid_indices.back()[0];
+    int c = grid_indices[pen_idx][(j + 1) % grid_indices[pen_idx].size()];
+    facets.emplace_back(a,b,c);
+  }
+
+  vector<shared_ptr<Mesh>> meshes = { make_shared<Mesh>(facets, positions) };
+  shared_ptr<graphics::Geometry> sphere = make_shared<Geometry>(meshes);
+  return sphere;
 }
 
 cv::Mat make_random(cv::Size size) {
@@ -138,7 +226,7 @@ std::shared_ptr<Geometry> Geometry::create_terrain(int z_len, int x_len) {
   cv::Mat noise0 = make_value_noise(cv::Size(x_len,z_len), seed0);
   cv::Mat noise1 = make_value_noise(cv::Size(x_len,z_len), seed1);
   cv::Mat noise = noise0 + 0.03 * noise1;
-  
+
   vector<glm::vec3> positions;
   for(int z = 0; z < z_len; z++) {
     for(int x = 0; x < x_len; x++) {
@@ -147,7 +235,7 @@ std::shared_ptr<Geometry> Geometry::create_terrain(int z_len, int x_len) {
       positions.push_back(glm::vec3(x_val,noise.at<float>(z,x),z_val));
     }
   }
-  
+
   vector<Facet> facets;
   auto find_idx = [](int z, int x, int X) -> int { return  X * z + x; };
   for(int z = 0; z < z_len - 1; z++) {
@@ -165,173 +253,8 @@ std::shared_ptr<Geometry> Geometry::create_terrain(int z_len, int x_len) {
       facets.push_back(Facet(a, b, c));
     }
   }
-  return make_shared<Geometry>(facets, positions);
-}
 
-std::shared_ptr<Geometry> Geometry::create_box(GLfloat x, GLfloat y, GLfloat z) {
-  x = x/2;
-  y = y/2;
-  z = z/2;
-  
-  std::vector<glm::vec3> positions = {
-    glm::vec3(-x,y,z), glm::vec3(x,y,z),      //front top left 0, top right 1
-    glm::vec3(-x,-y,z), glm::vec3(x,-y,z),    //front low left 2, low right 3
-    glm::vec3(x,y,-z), glm::vec3(-x,y,-z),    //back top left 4, top right 5
-    glm::vec3(x,-y,-z), glm::vec3(-x,-y,-z),  //back low left 6, low right 7
-  };
-  
-  std::vector<std::vector<GLuint>> indices = {
-    {0,2,1, 1,2,3}, //front
-    {4,6,5, 5,6,7}, //back
-    {5,7,0, 0,7,2}, //left
-    {1,3,4, 4,3,6}, //right
-    {5,0,4, 4,0,1}, //top
-    {2,7,3, 3,7,6}  //bottom
-  };
-  
-  shared_ptr<Geometry> box = std::make_shared<Geometry>();
-  for(int i = 0; i < indices.size(); i++) {
-    // get sub indices
-    std::vector<GLuint> sub_indices = {0,2,1, 1,2,3};
-    // get sub positions
-    std::vector<glm::vec3> sub_positions;
-    sub_positions.push_back(positions[ indices[i][0] ]);
-    sub_positions.push_back(positions[ indices[i][2] ]);
-    sub_positions.push_back(positions[ indices[i][1] ]);
-    sub_positions.push_back(positions[ indices[i][5] ]);
-    
-    auto face_geometry = Geometry(sub_positions,sub_indices);
-    *box += face_geometry;
-  }
-  return box;
-}
-
-std::shared_ptr<Geometry> Geometry::create_flat_sphere(GLfloat r, int st, int sc) {
-  if(st < 2 || sc < 3)
-    throw runtime_error("cannot have < 2 stacks or < 3 slices");
-  std::vector<glm::vec3> positions;
-  std::vector<GLint> indices;
-  
-  vector<vector<glm::vec3>> temp(st + 1, {});
-  //top section
-  temp.front().emplace_back(0,r,0);
-  //mid sections
-  float per_st_angle = 180.0 / st;
-  float per_sc_angle = 360.0 / sc;
-  for(int i = 1; i < st; i++) {
-    float st_angle = 90.0 - (per_st_angle * i);
-    float sc_r = r * cos(glm::radians(st_angle));
-    float y = r * sin(glm::radians(st_angle));
-    for(int j = 0; j < sc; j++) {
-      float sc_angle = per_sc_angle * j;
-      float x = sc_r * sin(glm::radians(sc_angle));
-      float z = sc_r * cos(glm::radians(sc_angle));
-      
-      temp[i].emplace_back(x,y,z);
-    }
-  }
-  //low section
-  temp.back().emplace_back(0,-r,0);
-  
-  //assign vertices and indices
-  //top
-  for(int j = 0; j < temp[1].size(); j++) {
-    positions.push_back(temp[0][0]);
-    positions.push_back(temp[1][j]);
-    positions.push_back(temp[1][(j + 1) % temp[1].size()]);
-  }
-  
-  //mid
-  for(int i = 1; i < temp.size() - 2; i++) {
-    for(int j = 0; j < temp[i].size(); j++) {
-      positions.push_back(temp[i][j]);
-      positions.push_back(temp[i + 1][j]);
-      positions.push_back(temp[i + 1][(j + 1) % temp[i].size()]);
-      
-      positions.push_back(temp[i][j]);
-      positions.push_back(temp[i + 1][(j + 1) % temp[i].size()]);
-      positions.push_back(temp[i][(j + 1) % temp[i].size()]);
-    }
-  }
-  
-  //bottom
-  int pen_idx = temp.size() - 2;
-  for(int j = 0; j < temp[pen_idx].size(); j++) {
-    positions.push_back(temp[pen_idx][j]);
-    positions.push_back(temp.back()[0]);
-    positions.push_back(temp[pen_idx][(j + 1) % temp[pen_idx].size()]);
-  }
-  
-  shared_ptr<graphics::Geometry> sphere = make_shared<Geometry>(positions);
-  return sphere;
-}
-
-std::shared_ptr<Geometry> Geometry::create_smooth_sphere(GLfloat r, int st, int sc) {
-  if(st < 2 || sc < 3)
-    throw runtime_error("cannot have < 2 stacks or < 3 slices");
-  std::vector<glm::vec3> positions;
-  std::vector<GLint> indices;
-  
-  vector<vector<int>> grid_indices(st + 1, {}); //position indices of grid points
-  int accum = 0;
-  
-  //top section
-  grid_indices.front().push_back(accum++);
-  positions.emplace_back(0,r,0);
-  //mid sections
-  float per_st_angle = 180.0 / st;
-  float per_sc_angle = 360.0 / sc;
-  for(int i = 1; i < st; i++) {
-    float st_angle = 90.0 - (per_st_angle * i);
-    float sc_r = r * cos(glm::radians(st_angle));
-    float y = r * sin(glm::radians(st_angle));
-    for(int j = 0; j < sc; j++) {
-      float sc_angle = per_sc_angle * j;
-      float x = sc_r * sin(glm::radians(sc_angle));
-      float z = sc_r * cos(glm::radians(sc_angle));
-      
-      positions.emplace_back(x,y,z);
-      grid_indices[i].push_back(accum++);
-    }
-  }
-  //low section
-  grid_indices.back().push_back(accum++);
-  positions.emplace_back(0,-r,0);
-  
-  //assign vertices and indices
-  vector<Facet> facets;
-  //top
-  for(int j = 0; j < grid_indices[1].size(); j++) {
-    int a = grid_indices[0][0];
-    int b = grid_indices[1][j];
-    int c = grid_indices[1][(j + 1) % grid_indices[1].size()];
-    facets.emplace_back(a,b,c);
-  }
-  
-  //mid
-  for(int i = 1; i < grid_indices.size() - 2; i++) {
-    for(int j = 0; j < grid_indices[i].size(); j++) {
-      int a = grid_indices[i][j];
-      int b = grid_indices[i + 1][j];
-      int c = grid_indices[i + 1][(j + 1) % grid_indices[i].size()];
-      facets.emplace_back(a,b,c);
-      
-      int d = grid_indices[i][j];
-      int e = grid_indices[i + 1][(j + 1) % grid_indices[i].size()];
-      int f = grid_indices[i][(j + 1) % grid_indices[i].size()];
-      facets.emplace_back(d,e,f);
-    }
-  }
-  
-  //bottom
-  int pen_idx = grid_indices.size() - 2;
-  for(int j = 0; j < grid_indices[pen_idx].size(); j++) {
-    int a = grid_indices[pen_idx][j];
-    int b = grid_indices.back()[0];
-    int c = grid_indices[pen_idx][(j + 1) % grid_indices[pen_idx].size()];
-    facets.emplace_back(a,b,c);
-  }
-  
-  shared_ptr<graphics::Geometry> sphere = make_shared<Geometry>(facets, positions);
-  return sphere;
+  vector<shared_ptr<Mesh>> meshes = { make_shared<Mesh>(facets, positions) };
+  shared_ptr<graphics::Geometry> terrain_geometry = make_shared<Geometry>(meshes);
+  return terrain_geometry;
 }
